add review component
